fix(all): revert to building all files
